+++ src/tooling/piral-cli/src/common/npm.test.ts
@@ -113,7 +113,8 @@ describe('npm Module', () => {
 
   it('dissects a fully qualified name with latest correctly', async () => {
     wrongCase = false;
-    const [name, version, hadVersion, type] = await dissectPackageName(process.cwd(), 'foo@latest');
+    const client = await determineNpmClient(process.cwd());
+    const [name, version, hadVersion, type] = await dissectPackageName(process.cwd(), 'foo@latest', client);
     expect(hadVersion).toBe(true);
     expect(version).toBe('latest');
     expect(name).toBe('foo');
@@ -122,7 +123,8 @@ describe('npm Module', () => {
 
   it('dissects a fully qualified name  with a specific version correctly', async () => {
     wrongCase = false;
-    const [name, version, hadVersion, type] = await dissectPackageName(process.cwd(), 'foo@1.2.3');
+    const client = await determineNpmClient(process.cwd());
+    const [name, version, hadVersion, type] = await dissectPackageName(process.cwd(), 'foo@1.2.3', client);
     expect(hadVersion).toBe(true);
     expect(version).toBe('1.2.3');
     expect(name).toBe('foo');
@@ -131,7 +133,8 @@ describe('npm Module', () => {
 
   it('dissects a simple name correctly', async () => {
     wrongCase = false;
-    const [name, version, hadVersion, type] = await dissectPackageName(process.cwd(), 'foo');
+    const client = await determineNpmClient(process.cwd());
+    const [name, version, hadVersion, type] = await dissectPackageName(process.cwd(), 'foo', client);
     expect(hadVersion).toBe(false);
     expect(version).toBe('latest');
     expect(name).toBe('foo');
@@ -140,7 +143,8 @@ describe('npm Module', () => {
 
   it('dissects a relative file name correctly', async () => {
     wrongCase = false;
-    const [name, version, hadVersion, type] = await dissectPackageName('/home/yolo', '../foo/bar');
+    const client = await determineNpmClient(process.cwd());
+    const [name, version, hadVersion, type] = await dissectPackageName('/home/yolo', '../foo/bar', client);
     expect(hadVersion).toBe(false);
     expect(version).toBe('latest');
     expect(name).toBe(resolve('/home/yolo', '../foo/bar'));
@@ -149,7 +153,8 @@ describe('npm Module', () => {
 
   it('dissects an absolute file name correctly', async () => {
     wrongCase = false;
-    const [name, version, hadVersion, type] = await dissectPackageName('/home/yolo', '/foo/bar');
+    const client = await determineNpmClient(process.cwd());
+    const [name, version, hadVersion, type] = await dissectPackageName('/home/yolo', '/foo/bar', client);
     expect(hadVersion).toBe(false);
     expect(version).toBe('latest');
     expect(name).toBe(resolve('/home/yolo', '/foo/bar'));
@@ -158,7 +163,8 @@ describe('npm Module', () => {
 
   it('dissects a git SSH repo name correctly', async () => {
     wrongCase = false;
-    const [name, version, hadVersion, type] = await dissectPackageName(process.cwd(), 'ssh://foo-bar.com/foo.git');
+    const client = await determineNpmClient(process.cwd());
+    const [name, version, hadVersion, type] = await dissectPackageName(process.cwd(), 'ssh://foo-bar.com/foo.git', client);
     expect(hadVersion).toBe(false);
     expect(version).toBe('latest');
     expect(name).toBe('git+ssh://foo-bar.com/foo.git');
@@ -167,7 +173,8 @@ describe('npm Module', () => {
 
   it('dissects a git HTTPS repo name correctly', async () => {
     wrongCase = false;
-    const [name, version, hadVersion, type] = await dissectPackageName(process.cwd(), 'https://foo-bar.com/foo.git');
+    const client = await determineNpmClient(process.cwd());
+    const [name, version, hadVersion, type] = await dissectPackageName(process.cwd(), 'https://foo-bar.com/foo.git', client);
     expect(hadVersion).toBe(false);
     expect(version).toBe('latest');
     expect(name).toBe('git+https://foo-bar.com/foo.git');
@@ -176,7 +183,8 @@ describe('npm Module', () => {
 
   it('dissects a scoped name correctly', async () => {
     wrongCase = false;
-    const [name, version, hadVersion, type] = await dissectPackageName(process.cwd(), '@foo/bar');
+    const client = await determineNpmClient(process.cwd());
+    const [name, version, hadVersion, type] = await dissectPackageName(process.cwd(), '@foo/bar', client);
     expect(hadVersion).toBe(false);
     expect(version).toBe('latest');
     expect(name).toBe('@foo/bar');
@@ -185,7 +193,8 @@ describe('npm Module', () => {
 
   it('dissects a scoped fully qualified name with latest correctly', async () => {
     wrongCase = false;
-    const [name, version, hadVersion, type] = await dissectPackageName(process.cwd(), '@foo/bar@latest');
+    const client = await determineNpmClient(process.cwd());
+    const [name, version, hadVersion, type] = await dissectPackageName(process.cwd(), '@foo/bar@latest', client);
     expect(hadVersion).toBe(true);
     expect(version).toBe('latest');
     expect(name).toBe('@foo/bar');
@@ -194,7 +203,8 @@ describe('npm Module', () => {
 
   it('dissects a scoped fully qualified name  with a specific version correctly', async () => {
     wrongCase = false;
-    const [name, version, hadVersion, type] = await dissectPackageName(process.cwd(), '@foo/bar@^1.x');
+    const client = await determineNpmClient(process.cwd());
+    const [name, version, hadVersion, type] = await dissectPackageName(process.cwd(), '@foo/bar@^1.x', client);
     expect(hadVersion).toBe(true);
     expect(version).toBe('^1.x');
     expect(name).toBe('@foo/bar');