+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 1.8.3 (tbd)
 
+- Fixed `piral-component` to work with React 19
 - Updated `breakpoints` to have flexible breakpoints
 - Added `breakpoints` parameter to `getAppInstance` options
 - Added `breakpoints` parameter to `createInstanceElement` function