+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Fixed `piral-component` to work with React 19
 - Fixed `importmap` entry in emulator for async centrally shared dependencies
+- Improved pilet scaffolding within a monorepo (#742)
 - Updated `breakpoints` to have flexible breakpoints
 - Added `breakpoints` parameter to `getAppInstance` options
 - Added `breakpoints` parameter to `createInstanceElement` function